[#9] refactor: 변경
